[Silver III] Title: 주유소, Time: 24 ms, Memory: 5220 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -9,25 +9,25 @@ int main(){
     cin.tie(NULL);
 
     int N;
-    vector<int> distance;
-    vector<int> oil_price;
-    int result = 0;
+    vector<long long> distance;
+    vector<long long> oil_price;
+    long long result = 0;
 
     cin >> N;
 
     for (int i = 0; i < N - 1; i++){
-        int input;
+        long long input;
         cin >> input;
         distance.push_back(input);
     }
 
     for (int i = 0; i < N; i++){
-        int input;
+        long long input;
         cin >> input;
         oil_price.push_back(input);
     }
 
-    int min_oil_price = oil_price[0];
+    long long min_oil_price = oil_price[0];
 
     for (int i = 0; i < N - 1; i++){
         if (oil_price[i] < min_oil_price){
